refactor: exporting components from single file
