+++ src/components/Footer/Footer.jsx
@@ -66,4 +66,4 @@ const Footer = () => {
   );
 };
 
-export default Footer;
+export { Footer };
