+++ src/components/Navbar/Navbar.jsx
@@ -52,4 +52,4 @@ const Navbar = () => {
   );
 };
 
-export default Navbar;
+export { Navbar };
