+++ src/components/index.js
@@ -1,0 +1,7 @@
+export { Navbar } from "./Navbar/Navbar";
+export { Footer } from "./Footer/Footer";
+export { Filter } from "./Filter/Filter";
+export { CartItems } from "./Cart/CartItems/CartItems";
+export { WishlistItems } from "./Wishlist/WishlistItems/WishlistItems";
+export { Products } from "./ProductListing/Products/Products";
+export { CartBill } from "./Cart/CartBill/CartBill";
