+++ src/pages/Cart/Cart.jsx
@@ -1,8 +1,7 @@
-import { CartBill } from "../../components/Cart/CartBill/CartBill";
-import { CartItems } from "../../components/Cart/CartItems/CartItems";
+import { CartBill, CartItems } from "./../../components/index";
 import "./Cart.css";
 
-const Cart = () => {
+export const Cart = () => {
   return (
     <main className="cart-wrapper">
       <h3 className="cart-heading">My Cart</h3>
@@ -14,5 +13,3 @@ const Cart = () => {
     </main>
   );
 };
-
-export default Cart;
