+++ src/pages/index.js
@@ -1,0 +1,7 @@
+export { Home } from "./Home/Home";
+export { Cart } from "./Cart/Cart";
+export { ProductListing } from "./ProductListing/ProductListing";
+export { Wishlist } from "./Wishlist/Wishlist";
+export { Login } from "./Authentication/Login/Login";
+export { Signup } from "./Authentication/Signup/Signup";
+export { Logout } from "./Authentication/Logout/Logout";
